Move i18n init and resource adding to rootlets
